stock

+++ nodeJs/class-notes/24-StockAPI/src/controllers/auth.js
@@ -5,6 +5,7 @@
 const User=require("../models/user")
 const Token=require("../models/token")
 const passwordEncrypt=require("../helpers/passwordEncrypt")
+const jwt=require("jsonwebtoken")
 
 module.exports={
     login: async(req,res)=>{
@@ -16,16 +17,18 @@ module.exports={
             if(user && user.password== passwordEncrypt(password)) {
 
                 if(user.is_active){
+                    if(!(process.env.TOKEN_MODEL=='JWT')){
                     // token üret
                     let tokenData=await Token.findOne({user_id: user._id})
                     if(!tokenData){
-
                         tokenData=await Token.create(
                             {
                                 user_id: user._id ,  
                                 token: passwordEncrypt(user._id+Date.now())                                
                             }
                         )
+
+
                         // ekstra
                         /*
                         const { randomUUID }=require("node:crypto") // random uniq bir değer üretiyor         
@@ -36,11 +39,26 @@ module.exports={
                             }
                         )*/
                     }
+
+                    
+
                     res.send({
                         error:false,
                         key:tokenData.token,
                         user
                     })
+
+                }else{
+                    const accessToken=jwt.sign(user.toJSON(), process.env.ACCESS_KEY,{expiresIn:'40m'})
+                    const refreshToken=jwt.sign({_id:user._id,password:user.password}, 
+                        process.env.REFRESH_KEY,{expiresIn:'2d'})
+
+                        res.send({
+                            error:false,
+                            bearer:{accessToken,refreshToken},
+                            user
+                        })
+                }
 
                 }else{
                     res.errorStatusCode=401
@@ -59,21 +77,101 @@ module.exports={
 
     },
     logout: async(req,res)=>{   
-        const auth=req.headers?.authorization || null // Token tokenKey
-        const tokenKey=auth ? auth.split(' ') : null // ['Token','tokenKey']
+        const auth=req.headers?.authorization || null // Token tokenKey or Bearer accessToken
+        const tokenKey=auth ? auth.split(' ') : null // ['Token','tokenKey'] ['Bearer','accessToken'] 
         
-        let result={}
-        if(tokenKey && tokenKey[0]=='Token'){
-            // console.log(tokenKey[1])
-            result=await Token.deleteOne({token:tokenKey[1]})
+        if(!(process.env.TOKEN_MODEL=='JWT')){
+            let result={}
+            if(tokenKey && tokenKey[0]=='Token'){
+                // console.log(tokenKey[1])
+                result=await Token.deleteOne({token:tokenKey[1]})
+            }
+            const messageLogout="Logout OK"
+           
+
+        }else{
+            const messageLogout="Dont need logout"
         }
         res.send({
             error:false,
-            message:"Logout OK",
+            message:messageLogout,
             result 
                         
         })
         
-    }
+    },
+    refresh: async(req,res)=>{
+
+         /*
+            #swagger.tags = ['Authentication']
+            #swagger.summary = 'JWT: Refresh'
+            #swagger.description = 'Refresh accessToken with refreshToken'
+            #swagger.parameters['body'] = {
+                in: 'body',
+                required: true,
+                schema: {
+                    bearer: {
+                        refresh: '...refreshToken...'
+                    }
+                }
+            }
+        */
+
+            const refreshToken = req.body?.bearer?.refresh
+
+            if (refreshToken) {
+    
+                const jwtData = await jwt.verify(refreshToken, process.env.REFRESH_KEY)
+    
+                if (jwtData) {
+    
+                    const { id, password } = jwtData
+    
+                    if (id && password) {
+    
+                        const user = await User.findOne({ _id: id })
+    
+                        if (user && user.password == password) {
+    
+                            if (user.isActive) {
+    
+                                // JWT AccessToken:
+                                const accessToken = jwt.sign(user.toJSON(), process.env.ACCESS_KEY, { expiresIn: '30m' })
+    
+                                res.status(200).send({
+                                    error: false,
+                                    bearer: {
+                                        access: accessToken
+                                    }
+                                })
+    
+                            } else {
+    
+                                res.errorStatusCode = 401
+                                throw new Error('This account is not active.')
+                            }
+                        } else {
+    
+                            res.errorStatusCode = 401
+                            throw new Error('Wrong id or password.')
+                        }
+                    } else {
+    
+                        res.errorStatusCode = 401
+                        throw new Error('There is not id and password in refreshToken.')
+                    }
+                } else {
+    
+                    res.errorStatusCode = 401
+                    throw new Error('sa')
+                }
+    
+            } else {
+                
+                res.errorStatusCode = 401
+                throw new Error('Please enter token.refresh')
+            }
+    
+    },
 
 }
